Add the WalkCtx type.
Sort of on the fence about using a Context because they're easy to
abuse, but considering the potential magnitude of the maps/features
we will be working with, I'd like to be able to implement timeouts and
cancellations. The Context will also shuttle around information specific
to this walk, i.e. ignore secret doors, only look for evil castles, etc.