First pass at implementing terrain generation.
Map: this type is just a nested array of floats. Raw output from
`Eval3`. This will eventually get dumped to a JSON blob or passed to a
visualization library.

MapGen: This will be a 'reusable' map generator. Users will be able to
modify the Stretch/Squish parameters ( and others I don't know are
relevant yet ).

Generate(): This is the real business. Uses parameters stored on the
MapGen struct to create a Map value which can be fed to a visualizer.